huge sync with defcon/circlecity updates
I did a lot of crazy stuff to make defcon work, and here it all is,
including the fixes that make it reasonably sane.  rfhs_predictable_phys
renames the phys based on capability so the right phys are always
running the right SSIDs.  This basically doesn't matter for the virtual
game, but it very much is required to run in meatspace with
heterogeneous hardware.